move dwarf code to dwarf.rs
